requirements/common.txt
sphinx_rtd_theme == 0.4.3
sphinx == 2.2.1
django-recaptcha == 3.0.0
-e git+https://github.com/urtzai/django-mobile.git#egg=django-mobile
django-pagination-bootstrap == 2.4.6
django-datetime-widget == 0.9.3
django-star-ratings == 0.9.2
-e git+https://github.com/urtzai/django-forum-app@c5730ed11e72e8b7918e5b98403b77c4ec1a3a0b#egg=django-forum-app
django-mptt == 0.15.0
Django == 4.2.13
mysqlclient == 2.2.0
pillow == 9.4.0
django-sortedm2m == 3.1.1
django-extensions == 3.1.5
django-bootstrap-form == 3.4
xmltodict == 0.13.0
django-model-utils == 4.1.1
django-photologue == 3.15.1
django-tinymce == 3.4.0
django-tagging == 0.5.0
django-registration-redux == 2.9
django-contrib-comments == 2.2.0
Mastodon.py == 1.8.0
pyTelegramBotAPI == 4.10.0
tweepy == 4.13.0
django-crispy-forms >= 1.10.0
django-guardian == 2.3.0
djangorestframework == 3.14.0
django-rest-auth == 0.9.5
django-cors-headers == 3.6.0
django-appconf == 1.0.5
django-autoslug == 1.9.8
django-filter == 2.4.0
gunicorn == 20.1.0
raven == 6.10.0
mcstatus == 10.0.2
django-debug-toolbar==3.2.4
django-embed-video==1.4.5
